Reorganized code into modules.
